fix(added-fucntion-to-check-whether-details-already-present): to check whether update, get and delete details are already present or not
BREAKING CHANGE:

+++ final/src/main/java/com/example/demo/Service/mongoservice.java
@@ -1,33 +1,56 @@
-package com.example.demo.Service;
+package com.example.demo.service;
 
-import com.example.demo.Entity.mongotable;
-import com.example.demo.Repository.mongorepository;
+import com.example.demo.entity.MongoTable;
+import com.example.demo.repository.MongodbRepository;
+import com.example.demo.service.InterfaceMongoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.util.Random;
 @Service
 
-public class mongoservice {
+public class MongoService implements InterfaceMongoService {
     @Autowired
-    private mongorepository repo;
+    private MongodbRepository repo;
 
-    public mongotable adddetails(mongotable mongo){
+    public MongoTable AddDetails(MongoTable mongo){
         Random rand = new Random();
         mongo.set_id(rand.nextInt(999));
         return repo.save(mongo);
     }
-    public mongotable getDetailsbyid(int _id){
+    public MongoTable GetDetailsbyId(int _id){
+        boolean exists = repo.existsById(_id);
+        System.out.println(exists);
+        if (!exists){
+            MongoTable details = new MongoTable();
+            details.setName("Details not");
+            details.setDepartment("Found");
+            return details;
+        }
+        
         return repo.findById(_id).get();
     }
-    public mongotable updatedetailsMongodb(mongotable mongo){
-        mongotable existingmongodetails = getDetailsbyid(mongo.get_id());
+    public MongoTable UpdateDetailsMongodb(MongoTable mongo){
+        boolean exists = repo.existsById(mongo.get_id());
+        System.out.println(exists);
+        if (!exists){
+            MongoTable data = new MongoTable();
+            data.setName("Details not");
+            data.setDepartment("Found");
+            return data;
+        }
+        MongoTable existingmongodetails = GetDetailsbyId(mongo.get_id());
          //repo.findById(mongo.get_id()).get();
          System.out.println(existingmongodetails.getName());
         existingmongodetails.setDepartment(mongo.getDepartment());
         existingmongodetails.setName(mongo.getName());
         return repo.save(existingmongodetails);
     }
-    public String deletemongodetails(int _id){
+    public String DeleteMongoDetails(int _id){
+        boolean exists = repo.existsById(_id);
+        System.out.println(exists);
+        if (!exists){
+            return "details not found";
+        }
         repo.deleteById(_id);
         return _id+" deletion successfull";
     }
